Share synproxy struct between linux/notlinux versions
